feat: splitted monolithic code into several files to enhance readability
